app/requirements.txt
Flask
PyYAML
requests
prometheus_client
